fix server-side socket listeners
The `close` and `error` events are not fired when the server shuts down.

The `message` event is the only event handled differently for the server-side socket and WebSocket client, as it's the handler for the message sent from the client to the server so it should have another event type `server::message`. Other events like `close` and `error` are fired and handled at the same time so no reason to have a special event here.

Make the `server::message` the only special event for the `on` handlers for the server-side sockets and use the same type of listeners for other types to handle all the events properly